+++ locales/CHANGELOG.md
@@ -1,4 +1,16 @@
-- `v20210902.1-pre` **
+- `v20210906.1` **Proposal management and other updates**
+
+  - Links embedded in translations are now passed in come from configuration
+  - Optional items
+  - Additional translatable content:
+    - Upload & manage proposal
+    - Authentication
+    - Download proposal
+    - Highlight proposal for an event organizing team
+    - Contact information
+
+
+- `v20210902.1-pre`
 
   - *Maintenance-version* - document version numbers are updated to reflect the **content** version
   - Revised fluent file formatting (using variables in *kebab-case*)